requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
rich>=13.0.0
pypdf>=4.0.0
python-docx>=1.0.0
tenacity>=8.0.0
sqlalchemy>=2.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
slowapi>=0.1.9

# Dev dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
